+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - OnHover over Node names.
 - Incorrect macro completion list when inside a string.
 - Parser now support `\` at end of line when calling a macro.
+- Parsing issue when line comment has `"` or `'`
 
 ## [0.4.5] - 2025-07-02